user注册，登录，个人中心，admin登录

+++ src/main/java/com/loki/server/service/UserService.java
@@ -7,7 +7,7 @@ import com.loki.server.model.User;
 public interface UserService {
 	void insert(User user);
 	boolean update(User user);
-	boolean delete(long id);
-	User findById(long id);
+	boolean delete(int id);
+	User findById(int id);
 	List<User> findAll();
 }
